Consolidate route registration tests into TwoFactorRoutesTest
