+++ tests/Feature/Users/EnableTwoFactorTest.php
@@ -4,8 +4,6 @@ namespace Tests\Feature\Users;
 
 use Illuminate\Support\Collection;
 use Illuminate\Support\Facades\Event;
-use Illuminate\Support\Facades\Route;
-use Orchestra\Testbench\Attributes\DefineEnvironment;
 use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\Attributes\Group;
 use PHPUnit\Framework\Attributes\Test;
@@ -195,19 +193,6 @@ class EnableTwoFactorTest extends TestCase
             ->assertRedirect('/cp/auth/confirm-password');
 
         $this->assertNull($user->two_factor_confirmed_at);
-    }
-
-    #[Test]
-    #[DefineEnvironment('disableTwoFactor')]
-    public function two_factor_enable_routes_are_not_registered_when_two_factor_is_disabled()
-    {
-        $this->assertFalse(Route::has('statamic.cp.users.two-factor.enable'));
-        $this->assertFalse(Route::has('statamic.cp.users.two-factor.confirm'));
-    }
-
-    protected function disableTwoFactor($app)
-    {
-        $app['config']->set('statamic.users.two_factor_enabled', false);
     }
 
     private function user()
