+++ tests/Feature/Users/TwoFactorChallengeTest.php
@@ -5,7 +5,6 @@ namespace Tests\Feature\Users;
 use Illuminate\Support\Collection;
 use Illuminate\Support\Facades\Event;
 use Illuminate\Support\Facades\Route;
-use Orchestra\Testbench\Attributes\DefineEnvironment;
 use PHPUnit\Framework\Attributes\Group;
 use PHPUnit\Framework\Attributes\Test;
 use PragmaRX\Google2FA\Google2FA;
@@ -226,18 +225,6 @@ class TwoFactorChallengeTest extends TestCase
         $this
             ->get('/requires-elevated-session')
             ->assertOk();
-    }
-
-    #[Test]
-    #[DefineEnvironment('disableTwoFactor')]
-    public function two_factor_challenge_routes_are_not_registered_when_two_factor_is_disabled()
-    {
-        $this->assertFalse(Route::has('statamic.cp.two-factor-challenge'));
-    }
-
-    protected function disableTwoFactor($app)
-    {
-        $app['config']->set('statamic.users.two_factor_enabled', false);
     }
 
     private function user()
