+++ tests/Feature/Users/TwoFactorRecoveryCodesTest.php
@@ -3,8 +3,6 @@
 namespace Tests\Feature\Users;
 
 use Illuminate\Support\Collection;
-use Illuminate\Support\Facades\Route;
-use Orchestra\Testbench\Attributes\DefineEnvironment;
 use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\Attributes\Group;
 use PHPUnit\Framework\Attributes\Test;
@@ -121,20 +119,6 @@ class TwoFactorRecoveryCodesTest extends TestCase
             ->actingAs($user)
             ->get(cp_route('users.two-factor.recovery-codes.download'))
             ->assertRedirect('/cp/auth/confirm-password');
-    }
-
-    #[Test]
-    #[DefineEnvironment('disableTwoFactor')]
-    public function two_factor_recovery_code_routes_are_not_registered_when_two_factor_is_disabled()
-    {
-        $this->assertFalse(Route::has('statamic.cp.users.two-factor.recovery-codes.show'));
-        $this->assertFalse(Route::has('statamic.cp.users.two-factor.recovery-codes.generate'));
-        $this->assertFalse(Route::has('statamic.cp.users.two-factor.recovery-codes.download'));
-    }
-
-    protected function disableTwoFactor($app)
-    {
-        $app['config']->set('statamic.users.two_factor_enabled', false);
     }
 
     private function user()
