+++ tests/Feature/Users/TwoFactorSetupTest.php
@@ -3,8 +3,6 @@
 namespace Tests\Feature\Users;
 
 use Illuminate\Support\Collection;
-use Illuminate\Support\Facades\Route;
-use Orchestra\Testbench\Attributes\DefineEnvironment;
 use PHPUnit\Framework\Attributes\Group;
 use PHPUnit\Framework\Attributes\Test;
 use Statamic\Auth\TwoFactor\RecoveryCode;
@@ -56,18 +54,6 @@ class TwoFactorSetupTest extends TestCase
                 'redirect' => 'https://evil.com',
             ]))
             ->assertRedirect(route('statamic.site'));
-    }
-
-    #[Test]
-    #[DefineEnvironment('disableTwoFactor')]
-    public function two_factor_setup_route_is_not_registered_when_two_factor_is_disabled()
-    {
-        $this->assertFalse(Route::has('statamic.cp.two-factor-setup'));
-    }
-
-    protected function disableTwoFactor($app)
-    {
-        $app['config']->set('statamic.users.two_factor_enabled', false);
     }
 
     private function user()
